Move font related types to src/app/fonts
