+++ src/app/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Aseprite
-# Copyright (C) 2018-2024  Igara Studio S.A.
+# Copyright (C) 2018-2025  Igara Studio S.A.
 # Copyright (C) 2001-2018  David Capello
 
 # Generate a ui::Widget for each widget in a XML file
@@ -98,11 +98,11 @@ add_library(app-lib ${generated_files})
 # These specific-platform files should be in an external library
 # (e.g. "base" or "os").
 if(WIN32)
-  target_sources(app-lib PRIVATE font_path_win.cpp)
+  target_sources(app-lib PRIVATE fonts/font_path_win.cpp)
 elseif(APPLE)
-  target_sources(app-lib PRIVATE font_path_osx.mm)
+  target_sources(app-lib PRIVATE fonts/font_path_osx.mm)
 else()
-  target_sources(app-lib PRIVATE font_path_unix.cpp)
+  target_sources(app-lib PRIVATE fonts/font_path_unix.cpp)
 endif()
 
 # This defines a specific webp decoding utility function for using
@@ -544,8 +544,9 @@ target_sources(app-lib PRIVATE
   file_system.cpp
   filename_formatter.cpp
   flatten.cpp
-  font_info.cpp
-  font_path.cpp
+  fonts/font_data.cpp
+  fonts/font_info.cpp
+  fonts/font_path.cpp
   gui_xml.cpp
   i18n/strings.cpp
   i18n/xml_translator.cpp
@@ -669,7 +670,6 @@ target_sources(app-lib PRIVATE
   ui/search_entry.cpp
   ui/select_accelerator.cpp
   ui/selection_mode_field.cpp
-  ui/skin/font_data.cpp
   ui/skin/skin_part.cpp
   ui/skin/skin_property.cpp
   ui/skin/skin_slider_property.cpp
