+++ src/app/commands/cmd_options.cpp
@@ -18,6 +18,7 @@
 #include "app/extensions.h"
 #include "app/file/file.h"
 #include "app/file_selector.h"
+#include "app/fonts/font_data.h"
 #include "app/i18n/strings.h"
 #include "app/ini_file.h"
 #include "app/launcher.h"
@@ -34,7 +35,6 @@
 #include "app/ui/rgbmap_algorithm_selector.h"
 #include "app/ui/sampling_selector.h"
 #include "app/ui/separator_in_view.h"
-#include "app/ui/skin/font_data.h"
 #include "app/ui/skin/skin_theme.h"
 #include "app/util/render_text.h"
 #include "base/convert_to.h"
