+++ src/app/fonts/font_info_tests.cpp
@@ -1,12 +1,12 @@
 // Aseprite
-// Copyright (c) 2024  Igara Studio S.A.
+// Copyright (c) 2024-2025  Igara Studio S.A.
 //
 // This program is distributed under the terms of
 // the End-User License Agreement for Aseprite.
 
 #include "tests/app_test.h"
 
-#include "app/font_info.h"
+#include "app/fonts/font_info.h"
 
 using namespace app;
 using namespace std::literals;
