+++ src/app/fonts/font_path.h
@@ -4,8 +4,8 @@
 // This program is distributed under the terms of
 // the End-User License Agreement for Aseprite.
 
-#ifndef APP_FONT_PATH_H_INCLUDED
-#define APP_FONT_PATH_H_INCLUDED
+#ifndef APP_FONTS_FONT_PATH_H_INCLUDED
+#define APP_FONTS_FONT_PATH_H_INCLUDED
 #pragma once
 
 #include "base/paths.h"
