+++ src/app/fonts/font_path_osx.mm
@@ -1,4 +1,5 @@
 // Aseprite
+// Copyright (C) 2025  Igara Studio S.A.
 // Copyright (C) 2017-2018  David Capello
 //
 // This program is distributed under the terms of
@@ -8,7 +9,7 @@
   #include "config.h"
 #endif
 
-#include "app/font_path.h"
+#include "app/fonts/font_path.h"
 
 #include "base/fs.h"
