+++ src/app/ui/skin/skin_theme.cpp
@@ -14,14 +14,14 @@
 #include "app/app.h"
 #include "app/console.h"
 #include "app/extensions.h"
-#include "app/font_info.h"
-#include "app/font_path.h"
+#include "app/fonts/font_data.h"
+#include "app/fonts/font_info.h"
+#include "app/fonts/font_path.h"
 #include "app/modules/gui.h"
 #include "app/pref/preferences.h"
 #include "app/resource_finder.h"
 #include "app/ui/app_menuitem.h"
 #include "app/ui/keyboard_shortcuts.h"
-#include "app/ui/skin/font_data.h"
 #include "app/ui/skin/skin_property.h"
 #include "app/ui/skin/skin_slider_property.h"
 #include "app/util/render_text.h"
